Problem: -------- From MySQL 8.0.21 release, redo logs can be disabled and enabled dynamically using ALTER INSTANCE [ENABLE/DISABLE] REDO_LOG. (This is meant only for initial data loading stages)
PXB relies on replaying redo logs to get consistent backups.

Fix:
----
Detect whether REDOLOG is disabled or not and error out as follows

1. Redo is disabled before backup                              : Backup failure
2. Redo is disabled and enabled before backup                  : Successful backup
3. Redo is enabled at start of backup, but disabled before end
 of backup (applies to both full and incremental backups)     : Backup failure
4. Redo is disabled and enabled again between Full and
Incremental backup  				              : Successful backup

Special case:
-------------

If redo is enabled at startup of backup but during the process of backup it is
disabled and enabled before the backup completes, it will result in inconsistent
backups. PXB may not error out in this scenario.

Users are advised to use --lock-ddl=ON with PXB to avoid users disabling and
enabling redo logs "during the backups".