PXB-2865 Failed to copy dymainc metadata https://jira.percona.com/browse/PXB-2865
We have the logic of skipping dynamic metadata on the apply-log phase
and relying on MySQL.ibd to copy those changes. see PXB-2180
The issue comes if there are change that are not flushed to mysql.ibd
before the Final incremental backup.
Final Incremental prepare will skip applying of DYNAMIC_METADATA
redo and mysql.ibd will not have those changes.

PXB does not allow final prepare on backup prepared already with
apply-log-only